+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Update to objectbox-c 0.10.0.
 * Update to objectbox-android 2.7.1.
 * Update to objectbox-swift 1.4.0.
+* String startsWith and endsWith condition: removed unused `descending` parameter, add `caseSensitive` parameter.
+* String greaterThan/lessThan condition: `withEqual` is deprecated, use the greaterOrEqual/lessOrEqual condition instead.
 
 0.7.0 (2020-08-14)
 ------------------